app/src/androidTest/java/QuizViewModelTest.kt
import android.util.Log
import org.junit.Assert.*
import org.junit.Test
import androidx.lifecycle.SavedStateHandle
import com.example.hello.R

class QuizViewModelTest{
    @Test
    fun providesExceptedQuestionText() {
        val savedStateHandle = SavedStateHandle()
        val quizViewModel = QuizViewModel(savedStateHandle)
        assertEquals(R.string.question_australia, quizViewModel.currentQuestionText)
    }
    @Test
    fun wrapsAroundQuestionBank(){
        val savedStateHandle = SavedStateHandle(mapOf(current_index_key to 5))
        val quizViewModel = QuizViewModel(savedStateHandle)
        assertEquals(R.string.question_asia, quizViewModel.currentQuestionText)
        quizViewModel.moveToNext()
        assertEquals(R.string.question_australia, quizViewModel.currentQuestionText)

    }
}